+++ src/components/App.tsx
@@ -1,14 +1,9 @@
-import Main from "../pages/Main";
-import { SettingProps } from "../props/SettingProps";
+import Main from '../pages/Main';
+import SettingProps from '../props/SettingProps';
+
 
 type AppProps = {
-    settings: SettingProps
-}
+  settings: SettingProps;
+};
 
- const App = ( {settings}: AppProps) => {
-  return (
-    <Main settings = {settings}></Main>
-  );
-}
-
-export default App;
+export const App = ({settings}: AppProps) => (<Main settings = {settings}></Main>);
