+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './components/App';
-import { SettingProps } from './props/SettingProps';
+import { App } from './components/App';
+import SettingProps from './props/SettingProps';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
